challenges/misc/kiss-me-through-the-scope/src/requirements.txt
Pillow
flask-limiter
flask
